Move all AdvApi32 to desktop
#223
