+++ src/AdvApi32.Desktop/AdvApi32+SECURITY_INFORMATION.cs
