+++ src/AdvApi32.Desktop/AdvApi32+SERVICE_STATUS.cs
