+++ src/AdvApi32.Desktop/AdvApi32+SafeServiceHandle.cs
