+++ src/AdvApi32.Desktop/AdvApi32+ServiceAccess.cs
