+++ src/AdvApi32.Desktop/AdvApi32+ServiceControl.cs
