+++ src/AdvApi32.Desktop/AdvApi32+ServiceControlAction.cs
