+++ src/AdvApi32.Desktop/AdvApi32+ServiceControlActionType.cs
