+++ src/AdvApi32.Desktop/AdvApi32+ServiceDelayedAutoStartInfo.cs
