+++ src/AdvApi32.Desktop/AdvApi32+ServiceDescription.cs
