+++ src/AdvApi32.Desktop/AdvApi32+ServiceErrorControl.cs
