+++ src/AdvApi32.Desktop/AdvApi32+ServiceFailureActions.cs
