+++ src/AdvApi32.Desktop/AdvApi32+ServiceLaunchProtected.cs
