+++ src/AdvApi32.Desktop/AdvApi32+ServiceManagerAccess.cs
