+++ src/AdvApi32.Desktop/AdvApi32+ServicePreshutdownInfo.cs
