+++ src/AdvApi32.Desktop/AdvApi32+ServiceRequiredPrivilegesInfo.cs
