+++ src/AdvApi32.Desktop/AdvApi32+ServiceSidInfo.cs
