+++ src/AdvApi32.Desktop/AdvApi32+ServiceSidType.cs
