+++ src/AdvApi32.Desktop/AdvApi32+ServiceStartType.cs
