+++ src/AdvApi32.Desktop/AdvApi32+ServiceState.cs
