+++ src/AdvApi32.Desktop/AdvApi32+ServiceType.cs
