+++ src/AdvApi32.Desktop/AdvApi32+TOKEN_ELEVATION_TYPE.cs
