+++ src/AdvApi32.Desktop/AdvApi32+TOKEN_INFORMATION_CLASS.cs
