+++ src/AdvApi32.Desktop/AdvApi32+TokenAccessRights.cs
