+++ src/AdvApi32.Desktop/AdvApi32.Desktop.csproj
@@ -64,17 +64,42 @@
     <Compile Include="AdvApi32.cs">
       <Generator>MSBuild:GenerateCodeFromAttributes</Generator>
     </Compile>
-    <Compile Include="AdvApi32.Helpers.cs" />
+    <Compile Include="AdvApi32.Helpers.cs">
+      <Generator>MSBuild:GenerateCodeFromAttributes</Generator>
+    </Compile>
     <Compile Include="AdvApi32.ReservedKeyHandles.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AdvApi32+ServiceTriggerDataType.cs" />
+    <Compile Include="AdvApi32+SafeServiceHandle.cs" />
+    <Compile Include="AdvApi32+SECURITY_INFORMATION.cs" />
+    <Compile Include="AdvApi32+ServiceAccess.cs" />
+    <Compile Include="AdvApi32+ServiceControl.cs" />
+    <Compile Include="AdvApi32+ServiceControlAction.cs" />
+    <Compile Include="AdvApi32+ServiceControlActionType.cs" />
+    <Compile Include="AdvApi32+ServiceDelayedAutoStartInfo.cs" />
+    <Compile Include="AdvApi32+ServiceDescription.cs" />
+    <Compile Include="AdvApi32+ServiceErrorControl.cs" />
+    <Compile Include="AdvApi32+ServiceFailureActions.cs" />
+    <Compile Include="AdvApi32+ServiceInfoLevel.cs" />
+    <Compile Include="AdvApi32+ServiceLaunchProtected.cs" />
+    <Compile Include="AdvApi32+ServiceManagerAccess.cs" />
+    <Compile Include="AdvApi32+ServicePreshutdownInfo.cs" />
+    <Compile Include="AdvApi32+ServiceRequiredPrivilegesInfo.cs" />
+    <Compile Include="AdvApi32+ServiceSidInfo.cs" />
+    <Compile Include="AdvApi32+ServiceSidType.cs" />
+    <Compile Include="AdvApi32+ServiceStartType.cs" />
+    <Compile Include="AdvApi32+ServiceState.cs" />
+    <Compile Include="AdvApi32+ServiceType.cs" />
+    <Compile Include="AdvApi32+SERVICE_STATUS.cs" />
+    <Compile Include="AdvApi32+TokenAccessRights.cs" />
+    <Compile Include="AdvApi32+TOKEN_ELEVATION_TYPE.cs" />
+    <Compile Include="AdvApi32+TOKEN_INFORMATION_CLASS.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="AdvApi32.exports.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
-  <Import Project="..\AdvApi32.Shared\AdvApi32.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))\EnlistmentInfo.targets" Condition=" '$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))' != '' " />
 </Project>
