--- src/AdvApi32.Shared/AdvApi32.Shared.projitems
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>15a8ded1-0f36-4b35-9973-22851da77544</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>PInvoke</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+SafeServiceHandle.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+SECURITY_INFORMATION.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceAccess.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceControl.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceControlAction.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceControlActionType.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceDelayedAutoStartInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceDescription.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceErrorControl.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceFailureActions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceInfoLevel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceLaunchProtected.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceManagerAccess.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServicePreshutdownInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceRequiredPrivilegesInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceSidInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceSidType.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceStartType.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceState.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+ServiceType.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+SERVICE_STATUS.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+TokenAccessRights.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+TOKEN_ELEVATION_TYPE.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32+TOKEN_INFORMATION_CLASS.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32.cs">
-      <Generator>MSBuild:GenerateCodeFromAttributes</Generator>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)AdvApi32.Helpers.cs" />
-  </ItemGroup>
-</Project>
