--- src/AdvApi32/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
-// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
-
-using System.Reflection;
-
-[assembly: AssemblyTitle("PInvoke.AdvApi32")]
-[assembly: AssemblyProduct("PInvoke")]
-[assembly: AssemblyCopyright("Copyright © Andrew Arnott 2015")]
