--- src/AdvApi32/project.json
@@ -1,8 +1,0 @@
-﻿{
-  "supports": { },
-  "dependencies": {
-  },
-  "frameworks": {
-    ".NETPortable,Version=v4.0,Profile=Profile92": { }
-  }
-}
